% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric.cluster.global.R
\name{metric.cluster.global}
\alias{metric.cluster.global}
\title{Global Clustering Coefficient}
\usage{
metric.cluster.global(Network, node.sample, triplet.sample)
}
\arguments{
\item{Network}{The input network.}

\item{node.sample}{The percentage of nodesto be selected - random sample (0,1).}

\item{triplet.sample}{The number of triplets to explore for each node - random triplets.}
}
\value{
A real constant.
}
\description{
Calculate the global clustering coefficient of a graph.
}
\details{
The global clustering coefficient measures the ratio of triples versus the total number of all possible triples in graph \emph{g}. \code{metric.cluster.global()} calculates the (estimated) global clustering coefficient of graph \emph{g} with a justified error.
}
\examples{
\dontrun{
x <-  net.erdos.renyi.gnp(1000, 0.01)
metric.cluster.global(x, 0.001,2)}
}
\author{
Xu Dong, Nazrul Shaikh.
}
\references{
Wasserman, Stanley, and Katherine Faust. Social network analysis: Methods and applications. Vol. 8. Cambridge university press, 1994.
}

