% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric.distance.apl.R
\name{metric.distance.apl}
\alias{metric.distance.apl}
\title{Average Path Length}
\usage{
metric.distance.apl(Network, probability, error, ncores = detectCores())
}
\arguments{
\item{Network}{The input network.}

\item{probability}{The confidence level probability.}

\item{error}{The sampling error}

\item{ncores}{The number of cores}
}
\value{
A real constant.
}
\description{
Calculate the average path length of a graph.
}
\details{
The average path length (APL) is the average shortest path lengths of all pairs of nodes in graph g. Calculates the (estimated) APL of graph \emph{g} with a justified error.
}
\examples{
\dontrun{
x <-  net.erdos.renyi.gnp(1000, 0.01)
metric.distance.apl(x, 0.01, 0.05)}
}
\author{
Luis Castro, Nazrul Shaikh.
}

