% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric.distance.diameter.R
\name{metric.distance.diameter}
\alias{metric.distance.diameter}
\title{Diameter}
\usage{
metric.distance.diameter(Network, probability = 0.95, error = 0.03,
  Cores = 1, full.apl = FALSE)
}
\arguments{
\item{Network}{The input network.}

\item{probability}{The confidence level probability}

\item{error}{The sampling error}

\item{Cores}{Number of cores to use in the computations}

\item{full.apl}{It will calculate the sampling version by default. If it is set to true, the population APL will be calculated and the rest of the parameters will be ignored.}
}
\value{
A real network
}
\description{
Calculate the diameter of a graph.
}
\details{
The diameter is the largest shortest path lengths of all pairs of nodes in graph \emph{Network}. \code{metric.distance.diameter} calculates the (estimated) diameter of graph \emph{Network} with a justified error.
}
\examples{
\dontrun{
##Default function
x <-  net.er.gnp(1000,0.01)
metric.distance.diameter(x)
##Population APL
metric.distance.diameter(x, full.apl=TRUE)
##Sampling at 99\% level with an error of 10\% using 5 cores
metric.distance.diameter(Network = x, probability=0.99, error=0.1, Cores=5)
}

}
\author{
Luis Castro, Nazrul Shaihk.
}
\references{
Dijkstra EW. A note on two problems in connexion with graphs:(numerische mathematik, _1 (1959), p 269-271). 1959.

Castro L, Shaikh N. Estimation of Average Path Lengths of Social Networks via Random Node Pair Sampling. Department of Industrial Engineering, University of Miami. 2016.
}

