% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric.distance.mpl.R
\name{metric.distance.mpl}
\alias{metric.distance.mpl}
\title{Median Path Length}
\usage{
metric.distance.mpl(Network, probability = 0.95, error = 0.03, Cores = 1,
  full.apl = FALSE)
}
\arguments{
\item{Network}{The input network.}

\item{probability}{The confidence level probability}

\item{error}{The sampling error}

\item{Cores}{Number of cores to use in the computations}

\item{full.apl}{It will calculate the sampling version by default. If it is set to true, the population APL will be calculated and the rest of the parameters will be ignored.}
}
\value{
A real network
}
\description{
Calculate the median path length of a graph.
}
\details{
The median path length (APL) is the median shortest path lengths of all pairs of nodes in graph \emph{Network}. \code{metric.distance.apl} calculates the population APL and estimated APL of graph g with a sampling error set by the user. The calculation uses a parallel load balancing approach, distributing jobs equally among the cores defined by the user.
}
\examples{
\dontrun{
##Default function
x <-  net.er.gnp(1000,0.01)
metric.distance.mpl(x)
##Population MPL
metric.distance.mpl(x, full.apl=TRUE)
##Sampling at 99\% level with an error of 10\% using 5 cores
metric.distance.mpl(Network = x, probability=0.99, error=0.1, Cores=5)
}

}
\author{
Luis Castro, Nazrul Shaihk.
}
\references{
Dijkstra EW. A note on two problems in connexion with graphs:(numerische mathematik, _1 (1959), p 269-271). 1959.

Castro L, Shaikh N. Estimation of Average Path Lengths of Social Networks via Random Node Pair Sampling. Department of Industrial Engineering, University of Miami. 2016.
}

